Addressed requested PR changes
