--- src/main/java/net/justonedev/kit/fileexplorer/FileChooserType.java
@@ -1,5 +1,0 @@
-package net.justonedev.kit.fileexplorer;
-
-public enum FileChooserType {
-    MONTH_PATH, TEX_PATH, PDF_PATH
-}
