+++ src/main/java/ui/fileexplorer/FileChooserType.java
@@ -1,0 +1,6 @@
+/* Licensed under MIT 2024. */
+package ui.fileexplorer;
+
+public enum FileChooserType {
+	MONTH_PATH, TEX_PATH, PDF_PATH
+}
